back-end/pyproject.toml
[project]
name = "edu-backend"
version = "0.1.0"
description = "Edu Marketplace Orders API"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "django-ninja>=1.3",
    "django-cors-headers>=4.4",
    "pydantic>=2.7",
    "granian[reload]>=1.6",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "httpx>=0.27",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["**/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "tests.py"]
addopts = "-x"
